Add files via upload
Team, I was having so much compile errors on this version. Tried debugging but not completed yet. Anyway, just to share my UML. 
Looks like Shawn's version in dev is looking better than my version, we can actually focus on that if you all agree. 
We obviously miscommunicated on how we should have approached but it looks like we are under controls!